# Upgrade benchviz to understand benchstat output.