benchstat is an upgraded benchcmp which can handle multiple benchmark runs and performs statistics on those runs.

https://godoc.org/rsc.io/benchstat