@ajstarks 2016-01-07
this change breaks benchcmp compatibility and does not handle bare ~.
working on a fix.