@mamyn0va 2020-07-24
> this change breaks benchcmp compatibility and does not handle bare ~.
> working on a fix.

Is there any update on benchviz update?
It would be great!

Thanks