@ajstarks 2020-07-27
I appears to me that the data you posted does not show a delta, and therefore benchviz has nothing to plot.