@mamyn0va 2020-07-31
Thanks, I misused `benchstat`.
I need to run my bench a lot more times (x10) to get significant p-results.